@@ -1,8 +1,21 @@
-const { webpack } = require("./helpers");
+import { webpack } from "./helpers";
+import RobotstxtPlugin from "../src/RobotstxtPlugin";
 
 describe("plugin", () => {
   it("should execute successfully", async () => {
     const stats = await webpack("entry.js");
+
+    const { warnings, errors, assets } = stats.compilation;
+
+    expect(warnings).toMatchSnapshot("warnings");
+    expect(errors).toMatchSnapshot("errors");
+    expect(assets["robots.txt"].source()).toMatchSnapshot();
+  });
+
+  it("should execute successfully #2", async () => {
+    const stats = await webpack("entry.js", {
+      plugins: [new RobotstxtPlugin()]
+    });
 
     const { warnings, errors, assets } = stats.compilation;
 
