@@ -1,1 +1,1 @@
-console.log("Hello world!"); // eslint-disable-line no-console
+console.log("Hello world!");
