@@ -6,9 +6,10 @@ module.exports = {
     "eslint --report-unused-disable-directives",
     "git add"
   ],
-  "*.{md,markdown,mdown,mkdn,mkd,mdwn,mkdown,ron}": [
+  "!(CHANGELOG).{md,markdown,mdown,mkdn,mkd,mdwn,mkdown,ron}": [
     "prettier --list-different",
     "remark -f -q",
     "git add"
-  ]
+  ],
+  "*.{yml,yaml}": ["prettier --list-different", "git add"]
 };
